requirements.txt
sanic==25.3.0
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.1.0
Pillow==11.1.0
httpx==0.28.1
